Updated parameters
Updated the `indir` and `outdir` parameters cause I don't like doing the extra step of clicking the underscore.

Also, updated the `sequence` flag into `illumina` or `ont` cause it looks better that way.

I am also considering renaming the pipeline to 'covcoi' because it sounds more legit and science-y. Naming the pipeline is the hardest task to do.